Iterative Explosions: Take 2 (#913)
* iterative explosions

* enable camera shake

* fixes

* microopt

* remove warn

* linter

* cleanup

* fix semi-infinite loop

* fix + r walls have high explosion block

* fuck

* fix log

* tweak explosion smoke

* fixes to throwing

* eh fuck it

* oops

* begin exploding/stop exploding

* fix proc ref

* eh, i like tyhis better

* devastating explosion human interaction fix

* fixup humans

* tweak human ex act

* more changes to human ex act

* fluid smoke

* https://github.com/tgstation/tgstation/pull/73555

* part of https://github.com/tgstation/tgstation/pull/76104, bother if perf is an issue

* fire prio stuff

* whoops

* fade in smoke

* paranoid

* boom boom shake the room

* bomb cap

* explosion throw refactor

* runtime

* eh

+++ code/modules/reagents/reagent_dispenser.dm
@@ -78,8 +78,9 @@
 		icon_state = "fuel_fools"
 
 /obj/structure/reagent_dispensers/fueltank/boom()
-	explosion(src, heavy_impact_range = 1, light_impact_range = 5, flame_range = 5)
+	var/turf/explode_turf = get_turf(src)
 	qdel(src)
+	explosion(explode_turf, heavy_impact_range = 1, light_impact_range = 5, flame_range = 5)
 
 /obj/structure/reagent_dispensers/fueltank/blob_act(obj/structure/blob/B)
 	boom()
@@ -130,6 +131,8 @@
 	tank_volume = 5000
 
 /obj/structure/reagent_dispensers/fueltank/large/boom()
+	if(QDELETED(src))
+		return
 	explosion(src, devastation_range = 1, heavy_impact_range = 2, light_impact_range = 7, flame_range = 12)
 	qdel(src)
